+++ test/utils/RouteUtils.spec.js
@@ -1,17 +1,17 @@
 import expect from 'expect';
-import * as UrlHelper from '../../scripts/helpers/UrlHelper';
+import * as RouteUtils from '../../scripts/utils/RouteUtils';
 
-describe('UrlHelper', () => {
+describe('RouteUtils', () => {
     describe('constructUrl', () => {
         it('should correctly construct a url', () => {
             const route = {path: ['songs'], query: {q: 'drake'}};
-            expect(UrlHelper.constructUrl(route)).toEqual('songs?q=drake');
+            expect(RouteUtils.constructUrl(route)).toEqual('songs?q=drake');
         });
     });
 
     describe('parseUrl', () => {
         it('should correctly parse a window hash', () => {
-            expect(UrlHelper.parseUrl('songs?q=drake')).toEqual({path: ['songs'], query: {q: 'drake'}});
+            expect(RouteUtils.parseUrl('songs?q=drake')).toEqual({path: ['songs'], query: {q: 'drake'}});
         });
     });
 });
